// src/app/tienda/models/usuarios.ts
export interface Usuarios {
    id: number;
    nombre: string;
    correo: string;
    edad: number;
    password: string;
    genero: string;
    estado: string;
    suscripcion: number;
    avatar: string;
    token:string;
    direccion:string;
    cp:number;
    provincia:string;
    pais:string;
}